PRA-48 Попытки в ssr [2]
