+++ packages/server/index.ts
@@ -14,8 +14,8 @@ const app = express();
 app.use(cors());
 const port = Number(process.env.SERVER_PORT) || 3001;
 
-app.get('/', (_, res) => {
-  const result = render();
+app.get('/', (req, res) => {
+  const result = render(req);
   const template = path.resolve(__dirname, '../client/dist/client/index.html');
   const htmlString = fs.readFileSync(template, 'utf-8');
   const newString = htmlString.replace('<!--ssr-outlet-->', result);
